Tidy imports and improve investment-market separation of concerns
